Added concurrent using std libraries
Added concurrent code using a very rough implementation of the
standard libraries.  Without this we have issues actually getting
the full use of the computer.  I"ll need to evaluate how we can
better utilze a thread pool so this way we can be more targetted
with the implmentation.

Initial results show 2x speed up on 4 cores