Added custom work sharing implementatation
Kind of ugly but performs better than messaging implementaitons.
we attempted in the past.  For each iteration we use a single
shared vector containing the random arrangement of nodeids
each of the nthreads we use then runs a loop while there are nodes
in this vector.  For each iteration of the loop it'll pop
off the last node and then run an update step using this node
for it's walk.